% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decode_detections}
\alias{decode_detections}
\title{decode_detections}
\usage{
decode_detections(detections_string)
}
\arguments{
\item{detections_string}{character, an endcoded string of semantic segmentation,
for example, "Gmt4AgoGbXB5L=="}
}
\value{
sf polygon
}
\description{
convert Mapillary object detection into sf polygons
}
\examples{
\donttest{
detection <- readLines(system.file('detection.txt', package = 'streetscape'))
streetscape::decode_detections(detection)
}
}
