% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_strview_by_id.R
\name{download_stv_by_id}
\alias{download_stv_by_id}
\title{download_stv_by_id}
\usage{
download_stv_by_id(
  ids = NULL,
  fields = c(),
  token = NULL,
  download = FALSE,
  size = 1,
  img_dir = NULL,
  parallel = FALSE
)
}
\arguments{
\item{ids}{numeric or character or vector.}

\item{fields}{vector, Additional fields of image metadata to retrieve.
Default includes: 'is_pano', 'height', 'width', 'computed_geometry',
'computed_altitude', and 'detections'.}

\item{token}{character, API token of Mapillary.}

\item{download}{logical. Whether to download the images.}

\item{size}{numeric. The size of images to download:
\itemize{
\item 1 - 256px wide thumbnail
\item 2 - 1024px wide thumbnail
\item 3 - 2048px wide thumbnail
\item 4 - Original size thumbnail
}
Default is 1.}

\item{img_dir}{character. The directory for downloading street view images.
Required if \code{download = TRUE}}

\item{parallel}{logical. Whether to parallelize the download process for speed.
The default is \code{FALSE}.}
}
\value{
A data.frame with metadata of the requested image IDs.
If \code{download = TRUE}, the selected image sizes will be saved in
\code{img_dir}.
}
\description{
Download Mapillary street view images and retrieve their metadata
by image id. The image url retrieved by \code{\link[=strview_searchByGeo]{strview_searchByGeo()}},
\code{\link[=strview_search_nnb]{strview_search_nnb()}}, or \code{\link[=strview_search_osm]{strview_search_osm()}} will expire after a period
of time.
}
\note{
More information about fields at
\url{https://www.mapillary.com/developer/api-documentation}
}
\examples{
\donttest{
ids <- c(527765141928031, 1437871829899771)
download_stv_by_id(ids = ids, token = 'token')
}

}
