% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_mvt}
\alias{install_mvt}
\title{install_mvt}
\usage{
install_mvt(envname = "r-mvt", method = "auto")
}
\arguments{
\item{envname}{The name, or full path, of the environment
in which Python packages are to be installed.}

\item{method}{character, indicating installation method.}
}
\value{
None
}
\description{
install_mvt is a wrapped function of py_install
in the reticulate package for installing the python package
mapbox_vector_tile, which will be installed in a virtual environment -
"r-mvt".
}
