% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdataframe-data.R
\docType{data}
\name{scdataframe}
\alias{scdataframe}
\title{streetscape dataframe}
\format{
An object of class \code{"StreetscapeDataFrame"}; see \code{\link[=see_streetscape_class]{see_streetscape_class()}}.
}
\usage{
data(scdataframe)
}
\description{
streetscape dataframe
}
\examples{
data(scdataframe)
}
\keyword{datasets}
